requirements.txt
click==8.1.3
dvc==2.41.1
feature-engine==1.5.2
matplotlib==3.6.3
nb_black==1.0.7
numpy==1.24.1
pandas==1.5.2
pydantic==1.10.4
PyYAML==6.0
scikit-learn==1.2.0
